+++ tests/utils.py
@@ -1,6 +1,6 @@
 # TODO test ft_clean, ft_free, ft_free_double_pointer
 
-from ctypes import CDLL, c_char_p, c_size_t, c_void_p, POINTER, Structure
+from ctypes import CDLL, c_char_p, c_int, c_size_t, c_void_p, POINTER, Structure
 
 INT_MAX = (1 << 31) - 1
 INT_MIN = ~INT_MAX
@@ -15,7 +15,7 @@ def same_sign(n1, n2):
 
 
 class t_dynamic_string(Structure):
-    _fields_ = [("content", c_char_p), ("length", c_size_t), ("capacity", c_size_t)]
+    _fields_ = [("content", c_char_p), ("length", c_int), ("capacity", c_int)]
 
 
 class t_list(Structure):
